// src/data/evolutions.gen.ts
export const EVOLUTION_MODES = {
  EVO_MODE_NORMAL: 0,
  EVO_MODE_TRADE: 1,
  EVO_MODE_ITEM_USE: 2,
  EVO_MODE_ITEM_CHECK: 3,
} as const;